had to move a function due to dependency stuff. fixed an error
